Authorization User, Product, Comment, Category

+++ app/Http/Controllers/UserController.php
@@ -15,7 +15,7 @@ class UserController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('checkLogin')->only('index', 'create', 'edit', 'update', 'destroy');
+        $this->middleware('checkLogin')->only('index', 'create', 'show', 'edit', 'update', 'destroy');
     }
     /**
      * Display a listing of the resource.
@@ -24,6 +24,7 @@ class UserController extends Controller
      */
     public function index()
     {
+        $this->authorize('view-any', User::class);
         $users = User::with('role')->orderBy('id', 'DESC')->paginate(5);
         return view('admin.users.index', compact('users'));
     }
@@ -35,6 +36,9 @@ class UserController extends Controller
      */
     public function create()
     {
+        if (Auth::check()) {
+            $this->authorize('create', User::class);
+        }
         $roles = Role::all();
         return view('admin.users.create', compact('roles'));
     }
@@ -47,6 +51,9 @@ class UserController extends Controller
      */
     public function store(UserStoreRequest $request)
     {
+        if (Auth::check()) {
+            $this->authorize('create', User::class);
+        }
         $user = new User;
         $user->fill($request->all());
         $user->password = Hash::make($request->password);
@@ -54,7 +61,7 @@ class UserController extends Controller
             if (Auth::check()) {
                 return redirect()->route('users.index')->with('notify', 'Thêm tài khoản thành công!');
             } else {
-                return redirect()->route('login')->with('notify','Bạn đã đăng ký thành công, vui lòng đăng nhập!');
+                return redirect()->route('login')->with('notify', 'Bạn đã đăng ký thành công, vui lòng đăng nhập!');
             }
         }
     }
@@ -78,6 +85,7 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        $this->authorize('update', $user);
         $roles = Role::all();
         return view('admin.users.edit', compact('user', 'roles'));
     }
@@ -91,8 +99,9 @@ class UserController extends Controller
      */
     public function update(UserStoreRequest $request, User $user)
     {
+        $this->authorize('update', $user);
         if ($user->update($request->all())) {
-            return redirect()->route('users.index')->with('notify','Sửa tài khoản thành công');
+            return redirect()->route('users.index')->with('notify', 'Sửa tài khoản thành công');
         }
     }
 
@@ -104,11 +113,12 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize('delete', $user);
         if ($user) {
             $comment = Comment::where('user_id', $user->id);
             $comment->delete();
             $user->delete();
         }
-        return redirect()->route('users.index')->with('notify','Xóa tài khoản thành công');
+        return redirect()->route('users.index')->with('notify', 'Xóa tài khoản thành công');
     }
 }
